+++ kimchi/src/circuits/constraints.rs
@@ -33,8 +33,6 @@ use std::sync::Arc;
 /// Flags for optional features in the constraint system
 #[derive(Copy, Clone, Serialize, Deserialize, Debug)]
 pub struct FeatureFlags {
-    /// ChaCha gates
-    pub chacha: bool,
     /// RangeCheck0 gate
     pub range_check0: bool,
     /// RangeCheck1 gate
@@ -104,10 +102,6 @@ pub struct ColumnEvaluations<F: PrimeField> {
     /// endoscalar multiplication selector over domain d8
     #[serde_as(as = "o1_utils::serialization::SerdeAs")]
     pub emul_selector8: E<F, D<F>>,
-
-    /// ChaCha selectors over domain d8
-    #[serde_as(as = "Option<[o1_utils::serialization::SerdeAs; 4]>")]
-    pub chacha_selectors8: Option<[E<F, D<F>>; 4]>,
 
     /// EC point addition selector over domain d8
     #[serde_as(as = "o1_utils::serialization::SerdeAs")]
@@ -483,44 +477,6 @@ impl<F: PrimeField + SquareRootField> ConstraintSystem<F> {
             .generic_selector
             .evaluate_over_domain_by_ref(self.domain.d4);
 
-        // chacha gate
-        let chacha_selectors8 = {
-            if !self.feature_flags.chacha {
-                None
-            } else {
-                Some([
-                    selector_polynomial(
-                        GateType::ChaCha0,
-                        &self.gates,
-                        &self.domain,
-                        &self.domain.d8,
-                        self.disable_gates_checks,
-                    ),
-                    selector_polynomial(
-                        GateType::ChaCha1,
-                        &self.gates,
-                        &self.domain,
-                        &self.domain.d8,
-                        self.disable_gates_checks,
-                    ),
-                    selector_polynomial(
-                        GateType::ChaCha2,
-                        &self.gates,
-                        &self.domain,
-                        &self.domain.d8,
-                        self.disable_gates_checks,
-                    ),
-                    selector_polynomial(
-                        GateType::ChaChaFinal,
-                        &self.gates,
-                        &self.domain,
-                        &self.domain.d8,
-                        self.disable_gates_checks,
-                    ),
-                ])
-            }
-        };
-
         // RangeCheck0 constraint selector polynomials
         let range_check0_selector8 = {
             if !self.feature_flags.range_check0 {
@@ -623,7 +579,6 @@ impl<F: PrimeField + SquareRootField> ConstraintSystem<F> {
             complete_add_selector4,
             mul_selector8,
             emul_selector8,
-            chacha_selectors8,
             endomul_scalar_selector8,
             range_check0_selector8,
             range_check1_selector8,
@@ -698,10 +653,39 @@ impl<F: PrimeField + SquareRootField> Builder<F> {
         // for some reason we need more than 1 gate for the circuit to work, see TODO below
         assert!(gates.len() > 1);
 
+        let lookup_features = LookupFeatures::from_gates(&gates, runtime_tables.is_some());
+
+        let num_lookups = {
+            let mut num_lookups: usize = lookup_tables
+                .iter()
+                .map(
+                    |LookupTable { data, id: _ }| {
+                        if data.is_empty() {
+                            0
+                        } else {
+                            data[0].len()
+                        }
+                    },
+                )
+                .sum();
+            for runtime_table in runtime_tables.iter() {
+                num_lookups += runtime_table.len();
+            }
+            let LookupFeatures { patterns, .. } = &lookup_features;
+            for pattern in patterns.into_iter() {
+                if let Some(gate_table) = pattern.table() {
+                    num_lookups += gate_table.table_size();
+                }
+            }
+            num_lookups
+        };
+
         //~ 2. Create a domain for the circuit. That is,
         //~    compute the smallest subgroup of the field that
         //~    has order greater or equal to `n + ZK_ROWS` elements.
-        let domain = EvaluationDomains::<F>::create(gates.len() + ZK_ROWS as usize)?;
+        let domain_size_lower_bound =
+            std::cmp::max(gates.len(), num_lookups + 1) + ZK_ROWS as usize;
+        let domain = EvaluationDomains::<F>::create(domain_size_lower_bound)?;
 
         assert!(domain.d1.size > ZK_ROWS);
 
@@ -717,10 +701,7 @@ impl<F: PrimeField + SquareRootField> Builder<F> {
             .collect();
         gates.append(&mut padding);
 
-        let lookup_features = LookupFeatures::from_gates(&gates, runtime_tables.is_some());
-
         let mut feature_flags = FeatureFlags {
-            chacha: false,
             range_check0: false,
             range_check1: false,
             lookup_features,
@@ -732,10 +713,6 @@ impl<F: PrimeField + SquareRootField> Builder<F> {
 
         for gate in &gates {
             match gate.typ {
-                GateType::ChaCha0
-                | GateType::ChaCha1
-                | GateType::ChaCha2
-                | GateType::ChaChaFinal => feature_flags.chacha = true,
                 GateType::RangeCheck0 => feature_flags.range_check0 = true,
                 GateType::RangeCheck1 => feature_flags.range_check1 = true,
                 GateType::ForeignFieldAdd => feature_flags.foreign_field_add = true,