+++ kimchi/src/error.rs
@@ -44,6 +44,9 @@ pub enum VerifyError {
     #[error("the previous challenges have an unexpected length (expected {0}, got {1})")]
     IncorrectPrevChallengesLength(usize, usize),
 
+    #[error("proof malformed: an evaluation was of the incorrect size (all evaluations are expected to be of length 1)")]
+    IncorrectEvaluationsLength,
+
     #[error("the opening proof failed to verify")]
     OpenProof,