+++ kimchi/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Upgrade to Rust 1.67.0
+- Remove unneeded ChaCha gates
 
 ## 0.1.0 (2023-02-06)